ASoC: sdm660: remove SAR switch
to be honest, i still don't understand what is this for, disable SAR
chip when turn on speaker and only for Indonesia?

currently, everytime speaker turn on, even for 1 secs, sar is disabled
and then enabled again within 1 sec, and it pointless. even max pro m2
has this disabled on latest kernel source.

Signed-off-by: Dede Dindin Qudsy <xtrymind@gmail.com>